Fix: css loop declaration
