--- Cliente/colonia_hlvs/src/screens/resident/invitations/single/singleInvitation.css
@@ -1,43 +1,0 @@
-@import url('../../../../assets/vars.css');
-
-.Left{
-    display: flex;
-    background-color: white;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-
-    font-family: var(--poppi), sans-serif;
-
-    .longText{
-        width: 40vh;
-        
-    }
-
-    .time_pickers{
-        width: 40vh;
-        display: flex;
-        justify-content: space-between;
-    
-        .time{
-            max-width: 18vh;
-        }
-    }
-
-    .input{
-        margin: 20px 0px;
-    }
-    
-}
-
-.Right{
-    background-color: black;
-
-}
-
-.father{
-    display: grid;
-    grid-template-columns: 50% 50%;
-    grid-template-rows: 100vh;
-    
-}
